Output ashmem memfd compatibility label as a file in the system partition
To aid with the transition from ashmem to memfd, all domains now have
permissions to have their tmpfs files (which is what memfds start off
as) relabeled to ashmem_compat_memfd. This label has similar access
control as the existing ashmem labels, which allows memfds to be used
without updating each required processes sepolicy.

To ensure that the policy is all in one place, add a rule to output the
label that libcutils should use when relabeling memfds.

Bug: 404959465
Bug: 404959459
Bug: 111903542
Change-Id: I5990e45777097f1e0f88d518c74522ae9043abd6
Signed-off-by: Isaac J. Manjarres <isaacmanjarres@google.com>