Add microdroid_manager.encrypted_store.status sysprop
This sysprop is used for encryptedstore binary to communicate to init
that encryptedstore was mounted so that init can run restorecon on it.

In the follow up patch this sysprop will also be used to notify payload
opted in delayed encryptedstore setup that the store is now ready to be
used.

Bug: 406258175
Test: atest MicrodroidTests
Flag: build.RELEASE_AVF_SUPPORT_LONG_RUNNING_VMS
Change-Id: I77189c3e6647b9fcff3628ed944c40d8291e0d22