Option to select Bluetooth certificate
Bluetooth is an optional mainline module that is using a shared uid.
It has some custom sepolicy rules that are associated to its
certificate.
When Bt is mainline, it can use MAINLINE_SEPOLICY_DEV_CERTIFICATES, but
when it is the from source prebuilt, it need to be
DEFAULT_SYSTEM_DEV_CERTIFICATE.

Due to the conditionality of this, Bluetooth need to have a custom
setting

Bug: 415283014
Bug: 415283014
Bug: 422180730
Bug: 409581884
Bug: 416451070
Bug: 408428360
Test: Flash a device with and without mainline module -> turn bt on
Flag: EXEMPT Provide build time config option for OEM
Change-Id: I3b76512a4fe372eaa3be91742391b2a65d75af52