Revert "Allow shell to use io_uring for testing purposes"
The use of userdebug_or_eng() is error prone and should be avoided
when possible. It introduces build-dependent differences in
behaviors which are historically not well tested. At Google, most
testing occurs on userdebug builds, not user builds. It also
violates the spirit go/android-userdebug-policy.

Instead of granting io_uring access for testing purposes, the
tests were modified to run as root. On debuggable builds, the
root user is not constrained by SELinux rules, and can do
whatever a test needs. There is no need to special case rules
for the shell user.

This reverts commit 7ed44b649cd8050874d893384690d08dbe16a8d0.

Test: atest IOUringSocketHandler_tests
Test: atest system/liburingutils
Bug: 406299670
Flag: EXEMPT test only change
Change-Id: I263f65d3102a854b4586766e0f05f105d9ab3784