Allow system_server to read sysfs_dm
PackageCacher manages its cache based on package filename. When an APEX
is mounted from a device-mapper block device, it creates a cache with a
block device filename (e.g. dm-17), which is not stable. The number can
change even if the APEX itsef is not changed.

Instead, PackageCacher can use the device mapper name for its cache key
if a given package is a device mapper block device because the device
name is stable across reboots. The device name can be read from sysfs
(e.g. /sys/block/dm-17/dm/name).

This change allows system_server to read the sysfs file for
device-mapper device name.

Bug: 427589794
Test: adb install <APEX> && reboot
Test: adb shell ls /data/system/package_cache/{hash}
Change-Id: I74403e97fc2ad3ac8bbb9d039f1ddf9b93f57832