shell: remove appdomain_tmpfs:file execute_no_trans
The use of memfd descriptors in an fexec() context
has similarities to https://g-issues.chromium.org/issues/40059061
which has been reported through ChromeOS VRP. As exec or fexec
on a memfd descriptor is not used, remove the execute_no_trans
permission which supports it.

Update the comments to indicate who is still using appdomain_tmpfs
functionality.

Bug: 425672336
Test: compiles and boots
Flag: EXEMPT TEST_ONLY
Change-Id: Id9923334ae3701795e989b9ad17f0d44e2fb004d