Fix odsign_e2e_tests
With Id67b139400f22e2dcd3b635ba74b6dc2755f1269, libvmclient calls isatty
and tcsetpgrp on stdout. This however broke odsign as explained below:

odsign forks compos_verify, which in turn forks virtmgr. When
compos_verify is forked, odsign uses logwrap_fork_execvp so that stdout
is redirected to logd. This internally creates a pty and label it as
odsign_devpts.

compos_verify uses libvmclient to spawn virtmgr. In doing so, the
library calls isatty and tcsetpgrp on stdout. This triggers ioctl on
odsign_devpts which is not allowed for compos_verify.

This change fixes the breakage by allowing the ioctl permission to
compos_verify.

Bug: 424226333
Test: run odsign_e2e_tests using ABTD
Flag: EXEMPT bug fix
Change-Id: I40f7df3a1bf1e5fe5e3467d4b4442a26cb376ef8