sepolicy: Add policy for userspace AOA
This adds the necessary labels for the userspace implementation of USB
Android Open Accessory protocol, and grants system_server permission to
relabel from functionfs to usbaccessory_device.

When userspace mounts a directory as functionfs, the kernel creates the
directory with a single file, ep0 in it. When this mount occurs, genfs
rules apply u:object_r:functionfs:s0 to this ep0 file. Userspace then
writes descriptors to ep0, and the kernel interprets the descriptors,
and for each endpoint described, creates an epX file, starting with ep1.

In the userspace AOA case, we write descriptors for two endpoints,
creating ep1 and ep2. Then, userspace will call restorecon on ep1 and
ep2 to apply the u:object_r:usbaccessory_device:s0 label to them. This
label allows app access to the endpoints, and mirrors the existing
/dev/usb_accessory file which currently has the usbaccessory_device
label.

We want ep0 to remain labeled as functionfs, as apps do not need access
to it.

Bug: 407985367
Bug: 424156976
Test: Manual, verify /dev/usb-ffs/aoa/ files are labeled correctly.
Change-Id: I43332a871ebb9d65fb0cb304ff486c4adc658345
Signed-off-by: Neill Kapron <nkapron@google.com>